Add api keys, authorization and management
We want to limit who can send a request, so we will use API keys.
Instead of having one for everyone it's basically just as easy to add
some basic multi-key management (with some very minor overhead for a db
call).

This PR adds two things:
1. A second required key for requests `api_key` a string in the JSON
object. If you don't have that you just get a 403 error, which is also
the same if you have a wrong API key.
2. Three new rake tasks to manage this they are:
    - `keys:list` List all the keys in the system
    - `keys:new[name goes here]` Create a new key with the given name
    - `keys:delete[id]` Delete a key with the given id